Add Wifi Hotspot functionality (#122)
This PR creates a passwordless WIFI Hotspot, when the camera has a wifi stick plugged in.

ssid: "axiom"
password: None

The camea has the ip 10.42.0.1 and it will be a dhcp & dns server. the domain axiom.camera will be resolved to this address.

A captive portal funcionality is implemented as follows:
All other domains will be resolved to 1.1.1.1. and all traffic to all adresses (except for 10.42.0.1) on port 80 will be redirected to 10.42.0.1:80 where a webserver will redirect to axiom.camera 
This allows the user to enter any ip and see the cameras webinterface. Moreover most modern computers will show the camera's web ui imediately after connecting.